Move some compilers into Component
