+++ phpstan.neon
@@ -11,6 +11,7 @@ parameters:
 
     excludePaths:
         - %currentWorkingDirectory%/src/Bundle/Controller/*
+        - %currentWorkingDirectory%/src/Bundle/DependencyInjection/Compiler/*
         - %currentWorkingDirectory%/src/Bundle/DependencyInjection/Configuration.php
         - %currentWorkingDirectory%/src/Bundle/DependencyInjection/PagerfantaConfiguration.php
         - %currentWorkingDirectory%/src/Bundle/DependencyInjection/Driver/Doctrine/DoctrineODMDriver.php
