+++ src/Bundle/DependencyInjection/Compiler/CsrfTokenManagerPass.php
@@ -13,21 +13,10 @@ declare(strict_types=1);
 
 namespace Sylius\Bundle\ResourceBundle\DependencyInjection\Compiler;
 
-use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
-use Symfony\Component\DependencyInjection\ContainerBuilder;
+\class_exists(\Sylius\Component\Resource\Symfony\Bundle\DependencyInjection\Compiler\CsrfTokenManagerPass::class);
 
-/**
- * TODO Remove on sylius/resource-bundle 2.0
- */
-final class CsrfTokenManagerPass implements CompilerPassInterface
-{
-    public function process(ContainerBuilder $container): void
+if (false) {
+    final class CsrfTokenManagerPass extends \Sylius\Component\Resource\Symfony\Bundle\DependencyInjection\Compiler\CsrfTokenManagerPass
     {
-        if (!$container->hasDefinition('security.csrf.token_manager')) {
-            return;
-        }
-
-        $csrdTokenManagerDefinition = $container->getDefinition('security.csrf.token_manager');
-        $csrdTokenManagerDefinition->setPublic(true);
     }
 }
