+++ src/Bundle/DependencyInjection/Compiler/RegisterFormBuilderPass.php
@@ -13,28 +13,11 @@ declare(strict_types=1);
 
 namespace Sylius\Bundle\ResourceBundle\DependencyInjection\Compiler;
 
-use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
-use Symfony\Component\DependencyInjection\ContainerBuilder;
-use Symfony\Component\DependencyInjection\Reference;
+\class_exists(\Sylius\Component\Resource\Symfony\Bundle\DependencyInjection\Compiler\RegisterFormBuilderPass::class);
 
-final class RegisterFormBuilderPass implements CompilerPassInterface
-{
-    public function process(ContainerBuilder $container): void
+if (false) {
+    final class RegisterFormBuilderPass extends \Sylius\Component\Resource\Symfony\Bundle\DependencyInjection\Compiler\RegisterFormBuilderPass
     {
-        if (!$container->hasDefinition('sylius.registry.form_builder')) {
-            return;
-        }
-
-        $registry = $container->findDefinition('sylius.registry.form_builder');
-
-        foreach ($container->findTaggedServiceIds('sylius.default_resource_form.builder') as $id => $attributes) {
-            foreach ($attributes as $attribute) {
-                if (!isset($attribute['type'])) {
-                    throw new \InvalidArgumentException('Tagged grid drivers needs to have "type" attribute.');
-                }
-
-                $registry->addMethodCall('register', [$attribute['type'], new Reference($id)]);
-            }
-        }
     }
 }
+
