+++ src/Bundle/DependencyInjection/Compiler/RegisterResourceRepositoryPass.php
@@ -13,30 +13,10 @@ declare(strict_types=1);
 
 namespace Sylius\Bundle\ResourceBundle\DependencyInjection\Compiler;
 
-use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
-use Symfony\Component\DependencyInjection\ContainerBuilder;
-use Symfony\Component\DependencyInjection\Reference;
+\class_exists(\Sylius\Component\Resource\Symfony\Bundle\DependencyInjection\Compiler\RegisterResourceRepositoryPass::class);
 
-final class RegisterResourceRepositoryPass implements CompilerPassInterface
-{
-    public function process(ContainerBuilder $container): void
+if (false) {
+    final class RegisterResourceRepositoryPass extends \Sylius\Component\Resource\Symfony\Bundle\DependencyInjection\Compiler\RegisterResourceRepositoryPass
     {
-        if (!$container->hasParameter('sylius.resources') || !$container->has('sylius.registry.resource_repository')) {
-            return;
-        }
-
-        /** @var array $resources */
-        $resources = $container->getParameter('sylius.resources');
-
-        $repositoryRegistry = $container->findDefinition('sylius.registry.resource_repository');
-
-        foreach ($resources as $alias => $configuration) {
-            [$applicationName, $resourceName] = explode('.', $alias, 2);
-            $repositoryId = sprintf('%s.repository.%s', $applicationName, $resourceName);
-
-            if ($container->has($repositoryId)) {
-                $repositoryRegistry->addMethodCall('register', [$alias, new Reference($repositoryId)]);
-            }
-        }
     }
 }
