+++ src/Bundle/DependencyInjection/Compiler/RegisterResourcesPass.php
@@ -13,40 +13,10 @@ declare(strict_types=1);
 
 namespace Sylius\Bundle\ResourceBundle\DependencyInjection\Compiler;
 
-use Sylius\Component\Resource\Model\ResourceInterface;
-use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
-use Symfony\Component\DependencyInjection\ContainerBuilder;
-use Symfony\Component\DependencyInjection\Exception\InvalidArgumentException;
+\class_exists(\Sylius\Component\Resource\Symfony\Bundle\DependencyInjection\Compiler\RegisterResourcesPass::class);
 
-final class RegisterResourcesPass implements CompilerPassInterface
-{
-    public function process(ContainerBuilder $container): void
+if (false) {
+    final class RegisterResourcesPass extends \Sylius\Component\Resource\Symfony\Bundle\DependencyInjection\Compiler\RegisterResourcesPass
     {
-        try {
-            /** @var array $resources */
-            $resources = $container->getParameter('sylius.resources');
-            $registry = $container->findDefinition('sylius.resource_registry');
-        } catch (InvalidArgumentException $exception) {
-            return;
-        }
-
-        foreach ($resources as $alias => $configuration) {
-            $this->validateSyliusResource($configuration['classes']['model']);
-            $registry->addMethodCall('addFromAliasAndConfiguration', [$alias, $configuration]);
-        }
-    }
-
-    private function validateSyliusResource(string $class): void
-    {
-        /** @var array $interfaces */
-        $interfaces = class_implements($class);
-
-        if (!in_array(ResourceInterface::class, $interfaces, true)) {
-            throw new InvalidArgumentException(sprintf(
-                'Class "%s" must implement "%s" to be registered as a Sylius resource.',
-                $class,
-                ResourceInterface::class,
-            ));
-        }
     }
 }
