+++ src/Component/Symfony/Bundle/DependencyInjection/Compiler/RegisterFormBuilderPass.php
@@ -1,0 +1,42 @@
+<?php
+
+/*
+ * This file is part of the Sylius package.
+ *
+ * (c) Sylius Sp. z o.o.
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace Sylius\Component\Resource\Symfony\Bundle\DependencyInjection\Compiler;
+
+use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\DependencyInjection\Reference;
+
+final class RegisterFormBuilderPass implements CompilerPassInterface
+{
+    public function process(ContainerBuilder $container): void
+    {
+        if (!$container->hasDefinition('sylius.registry.form_builder')) {
+            return;
+        }
+
+        $registry = $container->findDefinition('sylius.registry.form_builder');
+
+        foreach ($container->findTaggedServiceIds('sylius.default_resource_form.builder') as $id => $attributes) {
+            foreach ($attributes as $attribute) {
+                if (!isset($attribute['type'])) {
+                    throw new \InvalidArgumentException('Tagged grid drivers needs to have "type" attribute.');
+                }
+
+                $registry->addMethodCall('register', [$attribute['type'], new Reference($id)]);
+            }
+        }
+    }
+}
+
+\class_alias(RegisterFormBuilderPass::class, \Sylius\Bundle\ResourceBundle\DependencyInjection\Compiler\RegisterFormBuilderPass::class);
