+++ src/Component/Symfony/Bundle/DependencyInjection/Compiler/TwigPass.php
@@ -1,0 +1,35 @@
+<?php
+
+/*
+ * This file is part of the Sylius package.
+ *
+ * (c) Sylius Sp. z o.o.
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace Sylius\Component\Resource\Symfony\Bundle\DependencyInjection\Compiler;
+
+use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
+
+/**
+ * TODO Remove on sylius/resource-bundle 2.0
+ */
+final class TwigPass implements CompilerPassInterface
+{
+    public function process(ContainerBuilder $container): void
+    {
+        if (!$container->hasDefinition('twig')) {
+            return;
+        }
+
+        $twigDefinition = $container->getDefinition('twig');
+        $twigDefinition->setPublic(true);
+    }
+}
+
+\class_alias(TwigPass::class, \Sylius\Bundle\ResourceBundle\DependencyInjection\Compiler\TwigPass::class);
